refactor: validation helpers
